+++ tests/testthat/test_create_spp_aoh_data.R
@@ -87,7 +87,6 @@ test_that("simulated data (GDAL engine)", {
   skip_on_cran()
   skip_if_not_installed("gdalUtilities")
   skip_if_gdal_calc_not_available()
-  skip_on_os("mac")
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
   elevation_data <- terra::rast(
@@ -550,7 +549,6 @@ test_that("small range data (rasterize_touches, engine = gdal)", {
   skip_on_cran()
   skip_if_not_installed("gdalUtilities")
   skip_if_gdal_calc_not_available()
-  skip_on_os("mac")
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
   elevation_data <- terra::rast(